--- src/variables/global.js
@@ -1,11 +1,0 @@
-export const FILES_FOLDER = 'files';
-export const FS_ERROR = '(╯°□°）╯︵ ┻━┻ FS Operation failed.'
-export const GENERIC_ERROR = '(╯°□°）╯︵ ┻━┻  Operation failed.'
-export const STREAM_ERRROR = '(╯°□°）╯︵ ┻━┻ Stream operation failed.'
-export const INVALID_INPUT = '(╬ Ò﹏Ó) Invalid input.'
-export const PROMPT_MESSAGE = 'Welcome to the File manager,'
-export const EXIT_MESSAGE1 = 'Thank you for using File Manager, '
-export const EXIT_MESSAGE2 = ', goodbye! ᓚ₍ ^. .^₎'
-
-
-export const LOG_PATH = 'log.txt'
